Compute the bundle name once in the Makefile. Only in one place
Previously the bundle name was built both in the Makefile and in the Go
code. With this, we can now set it once. This is particularly useful if
the bundle name is not crc_hypervisor_version.
