+++ pkg/crc/constants/constants.go
@@ -71,20 +71,8 @@ func GetAdminHelperURL() string {
 	return adminHelperURLForOs[runtime.GOOS]
 }
 
-func defaultBundleForOs(bundleVersion string) map[string]string {
-	return map[string]string{
-		"darwin":  fmt.Sprintf("crc_hyperkit_%s.crcbundle", bundleVersion),
-		"linux":   fmt.Sprintf("crc_libvirt_%s.crcbundle", bundleVersion),
-		"windows": fmt.Sprintf("crc_hyperv_%s.crcbundle", bundleVersion),
-	}
-}
-
-func GetBundleFosOs(os, bundleVersion string) string {
-	return defaultBundleForOs(bundleVersion)[os]
-}
-
 func GetDefaultBundle() string {
-	return GetBundleFosOs(runtime.GOOS, version.GetBundleVersion())
+	return version.GetBundle()
 }
 
 var (
